Restructure and extend common build flags
This commit splits the old C++ flags up in C++ and C flags, and applies the ones that work in C, too, on C code as well as C++.

SQLite refuses to work with -ffast-math. We might drop the dependency entirely, though, as the old stats system is not needed any more.