Remove obsolete osxcross exception
Seems like it was all a misunderstanding in the first place, where a
file(GLOB ...) of CMake was interpreted differently on the platforms.

On macOS it appears to be possible to use case-insensitive file systems.
On Linux, where osxcross is used for testing, this is never possible.
Due to a spelling issue (SDLMain vs. SDLmain), the file was not found in
the first place, and the glob only returned two results. (Using a glob
was not necessary there anyway, to begin with).

In any case, main.m is not needed any more and can be removed. So can
the exception in CMake.