Revert "parkour physics tweak"
This reverts commit 3027c0244ac6aa604477d1d06f319a4296e90b3e.

This fixes #151

I'm reverting this commit because it causes the bug described in the
issue and reverting it fixes the bug.  The bug is fixed by making sure
we send the player's velocity to the `collide` function.  Inside the
collision code, the player's velocity is used to ignore collisions with
cube faces if the player is too deep inside the cube, as we can see in
this code line, part of definition of macro CHECKSIDE inside
engine/physics.cpp:

    if(d->type<ENT_CAMERA && dotval < 0 && dist < margin) continue; \

That code line is only executed if `dir` (the velocity of the player) is
non-zero, which this change makes sure that it will be.

Another possible fix could be to move the code line above out from the
if statement that checks if `dir` is non-zero, and also remove `dotval <
0` since `dotval` will be 0 if `dir` is 0, but the way I solve it in
this commit feels less risky.