Fix segmentation fault when changing map
This bug was introduced in commit 9b4d1185b902e830d6eabad6d001d715c24e7ae0

The reason that the segfault happens is that when one of these loops are
used with a std::vector, the size() returns an unsigned value and i will
therefore be unsigned and `i>=0` will always be true.  That will cause i
to underflow and when used as index to an array, that's how the segfault
happens.

The commit also adds a TODO comment about replacing those macros.