Change start command to take strings
This makes it possible to execute the start command like this:

    /start castle dm ffa duel

Most modes have an alias, like dm for deathmatch.  The old way of using
this command with numbers still works, but the difference is that the
command takes strings instead of numbers for mode and muts.  This
probably doesn't make any difference since types are implicitly
converted in cubescript.