Unify all assets into one single submodule
Instead of having all of the assets split into submodules for each of the
authors we now have one single submodule that contains them all.

This includes:

- Textures
- Sounds
- Models
- Maps

It is better to keep maps into their own, separate repo (they are a little
special compared to the rest, after all), which is included as a subtree
into the assets repo. Map contributions must still be done onto the maps
repo, so that changes can be later pulled into the assets.