Remove implicit fallthroughs
As of C++17, we can specify whether we want our switch cases to
fall-through or not (in presence of a break statement): specifying these
avoids unnecessary compiler warnings and makes the code clearer as to what
it is doing.

Specify the fallthrough attribute where needed and replace the comments
that specify its explicit usage.